requirements.txt
numpy
streamlit
Pillow
matplotlib
ultralytics
torch
torchvision
